Move failing tests to integration folder
