+++ packages/@sanity/cli/test/integration/commands/telemetry/__tests__/enable.test.ts
@@ -2,8 +2,8 @@ import {getCliToken} from '@sanity/cli-core'
 import {mockApi, testCommand} from '@sanity/cli-test'
 import {afterEach, beforeEach, describe, expect, test, vi} from 'vitest'
 
-import {fetchTelemetryConsent} from '../../../services/telemetry.js'
-import {Enable} from '../enable.js'
+import {fetchTelemetryConsent} from '../../../../../src/services/telemetry.js'
+import {Enable} from '../../../../../src/commands/telemetry/enable.js'
 
 vi.mock('@sanity/cli-core', async () => {
   const actual = await vi.importActual<typeof import('@sanity/cli-core')>('@sanity/cli-core')
@@ -14,8 +14,8 @@ vi.mock('@sanity/cli-core', async () => {
   }
 })
 
-vi.mock('../../../services/telemetry.js', async () => ({
-  ...(await vi.importActual('../../../services/telemetry.js')),
+vi.mock('../../../../../src/services/telemetry.js', async () => ({
+  ...(await vi.importActual('../../../../../src/services/telemetry.js')),
   fetchTelemetryConsent: vi.fn(),
 }))
